build: Add support for faking platform security patch level
SafetyNet will fail if the fake build fingerprints don't match the
platform security patch level, which is a problem for devices that are
no longer supported by the manufacturer.

To fix this issue, allow legacy products to define a fake security patch
level from their last official update and set the standard patch level
property to that. The real patch level is moved to another custom
property that is shown to the user, e.g. in Settings -> About.

Note that this requires the "core: Use real security patch level
property" commit in frameworks/base for users to see the correct patch
level in Settings.

Change-Id: Id09e1b52697e6791987e388551e6f7a6558775ba
Signed-off-by: spezi77 <spezi7713@gmx.net>
Signed-off-by: Joey Huab <joey@evolution-x.org>